requirements.txt
h5py>=3.8
numpy>=1.24
pywebview>=5.0
Pillow>=10.0
